chore: move inplayer types to a more suitable location and remove a redundant folder
